pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus"
version = "1.0.0"
description = "Autonomous hedge fund multi-agent trading system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "NEXUS Team"}]

dependencies = [
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "pyyaml>=6.0,<7",
    "python-dotenv>=1.0,<2",
    "httpx>=0.25,<1",
    "pandas>=2.1,<3",
    "numpy>=1.26,<2",
    "yfinance>=0.2.31,<1",
    "alpaca-py>=0.21,<1",
    "sqlalchemy>=2.0,<3",
    "psycopg2-binary>=2.9,<3",
    "alembic>=1.13,<2",
    "influxdb-client>=1.38,<2",
    "chromadb>=0.4,<1",
    "anthropic>=0.39,<1",
    "openai>=1.6,<2",
    "langchain>=0.1,<1",
    "langgraph>=0.0.40,<1",
    "langchain-anthropic>=0.1,<1",
    "langchain-openai>=0.0.5,<1",
    "langfuse>=2.7,<3",
    "wandb>=0.16,<1",
    "ta>=0.11,<1",
    "transformers>=4.36,<5",
    "torch>=2.1,<3",
    "rich>=13.7,<14",
    "click>=8.1,<9",
    "prometheus-client>=0.19,<1",
    "feedparser>=6.0,<7",
    "beautifulsoup4>=4.12,<5",
    "praw>=7.7,<8",
    "jinja2>=3.1,<4",
    "streamlit>=1.29,<2",
    "plotly>=5.18,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4.1,<6",
    "pytest-mock>=3.12,<4",
    "ruff>=0.1.9,<1",
    "mypy>=1.8,<2",
    "black>=23.12,<25",
    "pre-commit>=3.6,<4",
    "types-PyYAML>=6.0,<7",
    "types-requests>=2.31,<3",
]

[project.scripts]
nexus = "nexus.cli.main:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["nexus"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
